Last commit

+++ BasicCalculator/src/main/java/com/pluralsight/BasicCalculator.java
@@ -7,14 +7,14 @@ public class BasicCalculator {
        Scanner scanner = new Scanner(System.in);
        // Two float numbers
         System.out.println(" Enter the first number: ");
-        Float num1 = scanner.nextFloat();
+        float num1 = scanner.nextFloat();
 
         System.out.println(" Enter the Second number ");
-        Float num2 = scanner.nextFloat();
+        float num2 = scanner.nextFloat();
 
         // Possible outcomes
         System.out.println(" Enter one of the 4 operations: 1.Add, 2.Subtract, 3.Mulitply, 4.Divide ");
-        int choice = scanner.nextInt();
+        float choice = scanner.nextFloat();
 
         if (choice== 1) {
             System.out.println("Result: " + (num1 + num2));
